url: Introduce MakeRange(size_t, size_t)
This is helpful to reduce type casts.
This CL fixes ambiguous function call errors due to two MakeRange().

Bug: 350788890
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7123400
Owners-Override: Hayato Ito <hayato@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Auto-Submit: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540992}

+++ url/url_canon_host.cc
@@ -510,7 +510,7 @@ void DoHost(std::basic_string_view<CHAR> spec,
   // URL Standard: https://url.spec.whatwg.org/#host-parsing
 
   // Keep track of output's initial length, so we can rewind later.
-  const int output_begin = output.length();
+  const size_t output_begin = output.length();
 
   if (host.is_empty()) {
     // Empty hosts don't need anything.
